requirements.txt
requests
BeautifulSoup4
selenium
newspaper3k